updates
• Change name of setup fee to onetime fee[Pradeep]  (update)
• In the front page, next to price under the plans it should show "Per
Provider/Month"[Pradeep]  (Updated)
• When I made a product with a onetime fee of 2000(see data conversion)
and selected it, when I came to payment, it did not show on my invoice.
I did not choose setup fee for the product because i thought it was not
right. but if you change that to onetime fee then I can use that column
instead of messing with the recurring fee.  Question is why should we
have the option for the monthly fee to be one time or recurring if we
have a setup fee which is a onetime fee?
• Calculations shown in the confirmation page is not correct.  The
providers times plan amount minus discount is not correct. (it was
showing 3X150-35 as 520) I am guessing it was not taking the plan price
but the original price.[Pradeep]  (There is two setting. What I did is
that, if we providing discount type as “Price”, it is considered as flat
discount. And will reduce the discount amount from the total. If we set
discount in “%” then as per provider percentage discount will deduct.
[Pradeep] For example the Professional Plan,

Total Price is $185 and Discount in price is $35 so effective price is
$150
If choose two provider : $185 x 2 = $370 - $35 = $335.
If we set the percentage wise then will consider percentage in total
price as per provider selection.
• If there are no setup fees(onetime fee) can we not show that
column?[Pradeep]  (updated)
• Give hyperlink to popups the details of a product everywhere.[Pradeep]
(Provided)
• When i submit the back button once i am in the confirmation of payment
page then it messes up the data.  How can we stop this behavior.
sometimes people will go to the confirmation page and then decide to go
back and add another product and then come back it screws up everything.
[Pradeep] (Stopped the browser back feature on whole site)